fix: Set every path, merge every path
On windows systems, as of 3.1.0, while we didn't end up with an env that
had multiple competing and different /^PATH$/i environs, we could
sometimes pick the 'canonical' one in ways that broke other things.

Windows git-bash appears to sometimes have Path set to an empty string,
and PATH set to the actual %Path% from the parent system.  Some existing
tools update Path, but not PATH, or vice versa.

This caused tests to break for Lerna users, because Travis's lerna
bootstrap command could not find the 'npm' executable.

The solution here is to just read all PATH environs, and merge them
together.  Then, when we set the env for the child process, set them all
as well.